restructure to one class per file (#35)
