+++ src/main/scala/co/coinsmith/kafka/cryptocoin/StreamingActor.scala
@@ -1,0 +1,12 @@
+package co.coinsmith.kafka.cryptocoin
+
+import akka.actor.{Actor, Props}
+import co.coinsmith.kafka.cryptocoin.streaming.OKCoinStreamingActor
+
+
+class StreamingActor extends Actor {
+  val okcoin = context.actorOf(Props[OKCoinStreamingActor], "okcoin")
+  def receive = {
+    case _ =>
+  }
+}
